Refactor out ResourceBase from Purchases API.
